+++ lib/federation.js
@@ -37,6 +37,7 @@ const {
 const { validateSDL } = require('graphql/validation/validate')
 const compositionRules = require('./federation/compositionRules')
 const { MER_ERR_GQL_INVALID_SCHEMA, MER_ERR_GQL_GATEWAY_INVALID_SCHEMA } = require('./errors')
+const { hasExtensionDirective } = require('./util')
 
 const BASE_FEDERATION_TYPES = `
   scalar _Any
@@ -79,13 +80,6 @@ function removeExternalFields (typeDefinition) {
     ...typeDefinition,
     fields: typeDefinition.fields.filter(fieldDefinition => !fieldDefinition.directives.some(directive => directive.name.value === 'external'))
   }
-}
-
-function hasExtensionDirective (definition) {
-  if (!definition.directives) {
-    return false
-  }
-  return definition.directives.some(directive => directive.name.value === 'extends' || directive.name.value === 'requires')
 }
 
 function getStubTypes (schemaDefinitions, isGateway) {
