+++ lib/gateway/service-map.js
@@ -13,21 +13,7 @@ const { buildRequest, sendRequest } = require('./request')
 const SubscriptionClient = require('../subscription-client')
 const { MER_ERR_GQL_GATEWAY_INIT } = require('../errors')
 const buildFederationSchema = require('../federation')
-
-function hasDirective (directiveName, node) {
-  if (!node.directives || node.directives.length < 1) {
-    return false
-  }
-  return node.directives.some(directive => directive.name.value === directiveName)
-}
-
-function hasExensionDirective (node) {
-  if (!node.directives || node.directives.length < 1) {
-    return false
-  }
-  return node.directives.some(directive =>
-    directive.name.value === 'extends' || directive.name.value === 'requires')
-}
+const { hasDirective, hasExtensionDirective } = require('../util')
 
 function createFieldSet (existingSet, definition, filterFn = () => false) {
   const fieldsSet = existingSet || new Set()
@@ -50,7 +36,7 @@ function createTypeMap (schemaDefinition) {
   const extensionTypeMap = {}
 
   for (const definition of parsedSchema.definitions) {
-    const isTypeExtensionByDirective = hasExensionDirective(definition)
+    const isTypeExtensionByDirective = hasExtensionDirective(definition)
     /* istanbul ignore else we are only interested in type definition and type extension scenarios */
     if (isTypeDefinitionNode(definition) && !isTypeExtensionByDirective) {
       typeMap[definition.name.value] = createFieldSet(typeMap[definition.name.value], definition)
